test(380-04): match StallResilience gap-backfill keccak to frozen uint24 day width
test_stallSwapResume packed gap days 3/4/5 as uint32 in the keccak preimage; the
frozen _backfillGapDays packs them as uint24. Same width-mismatch class as the
VRFStallEdgeCases fixes. Frozen subject byte-untouched.

+++ test/fuzz/StallResilience.t.sol
@@ -104,16 +104,18 @@ contract StallResilience is DeployProtocol {
 
         // Verify gap day words are deterministic derivations of the resume VRF word.
         // _backfillGapDays is called BEFORE _applyDailyRng in rngGate.
-        // So it uses rngWordCurrent (the raw VRF word, pre-nudge).
-        uint256 expectedDay3 = uint256(keccak256(abi.encodePacked(resumeWord, uint32(3))));
+        // So it uses rngWordCurrent (the raw VRF word, pre-nudge). The frozen
+        // _backfillGapDays (AdvanceModule:1843-1845) packs the gap day as uint24 (its loop
+        // counter type), so the preimage day width is 3 bytes — match it with uint24.
+        uint256 expectedDay3 = uint256(keccak256(abi.encodePacked(resumeWord, uint24(3))));
         if (expectedDay3 == 0) expectedDay3 = 1;
         assertEq(game.rngWordForDay(3), expectedDay3, "Day 3 word is keccak256(vrfWord, 3)");
 
-        uint256 expectedDay4 = uint256(keccak256(abi.encodePacked(resumeWord, uint32(4))));
+        uint256 expectedDay4 = uint256(keccak256(abi.encodePacked(resumeWord, uint24(4))));
         if (expectedDay4 == 0) expectedDay4 = 1;
         assertEq(game.rngWordForDay(4), expectedDay4, "Day 4 word is keccak256(vrfWord, 4)");
 
-        uint256 expectedDay5 = uint256(keccak256(abi.encodePacked(resumeWord, uint32(5))));
+        uint256 expectedDay5 = uint256(keccak256(abi.encodePacked(resumeWord, uint24(5))));
         if (expectedDay5 == 0) expectedDay5 = 1;
         assertEq(game.rngWordForDay(5), expectedDay5, "Day 5 word is keccak256(vrfWord, 5)");
     }
